// src/core/enums/responses/response-code.enum.ts
// Response Error Code
export enum AppErrorCode {
  INTERNAL_SERVER_ERROR = '500',
  PRISMA_ERROR = '301',
  BAD_REQUEST = '400',
  UNAUTHORIZED = '401',
  FORBIDDEN = '403',
  NOT_FOUND = '404',
  HEADERS_INVALID = '405',
  VALIDATION_ERROR = '406',
  USER_STATUS_NOT_FOUND = '411',
  APPLICATION_NOT_FOUND = '421',
  USER_MODULE_NOT_FOUND = '431',
  USER_ROLE_NOT_FOUND = '441',
  USER_ACCOUNT_NOT_FOUND = '451',
  USER_ACCOUNT_PASSWORD_MUST_BE_DIFFERENT_WITH_NEW_ONE = '452',
  USER_ACCOUNT_CURRENT_PASSWORD_NOT_MATCH = '453',
}

export enum AppErrorCodeCustome {
  CUSTOME = '333',
}